### 2023-10-27 a515717
Apparently you need numbered affiliations even if there is only one author.
