### 2023-10-27 #5993 @Wrzlprmft
Some suggestions for reviewers:

- [Lucas Roberts](https://orcid.org/0000-0002-5249-0684) (involved in SciPy’s `combine_pvalues`)
- [Albert Steppi](https://orcid.org/0000-0001-5871-6245) (involved in SciPy’s `combine_pvalues`)
- [Micheal Dewey](https://orcid.org/0000-0002-7522-3677) (author of related R package)
- [Ozan Cinar](https://orcid.org/0000-0003-0329-1977) (author of related R package)
- [Wolfgang Viechtbauer](https://orcid.org/0000-0003-3463-4063) (author of related R package)
