### 2023-11-24 #5993 @vissarion
:wave: @rlucas7, @ozancinar, @steppi would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
