### 2023-11-28 #6096 @editorialbot
[REVIEW]: A Python module to combine p values arising from discrete tests
**Submitting author:** @Wrzlprmft (<a href="http://orcid.org/0000-0002-5472-7067">Gerrit Ansmann</a>)
**Repository:** https://github.com/BPSB/combine-p-values-discrete
**Branch with paper.md** (empty if default branch):
**Version:** 1.2.1
