### 2023-11-29 BPSB/combine-p-values-discrete#1 @mdhaber
Review of JOSS submission
Hi @Wrzlprmft; thanks for your submission to JOSS (https://github.com/openjournals/joss-reviews/issues/6096). As a scipy.stats maintainer and author of [Fast Resampling and Monte Carlo Methods in Python](https://joss.theoj.org/papers/10.21105/joss.05092.pdf), I was very interested.

Here are a few questions and comments from my review (so far).

- [x] 1. Can you respond to JOSS's criteria for "substantial scholarly effort", specifically "As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual."? It's unclear from the commit history whether that's satisfied, but commit histories don't tell the full story. (I know because I submitted a manuscript to JOSS for which this was questioned. Of course, I wouldn't have submitted if I didn't believe the criterion had been satisfied. But as a reviewer, I need to ask!)
- [x] 2. In the claim "The presented `module combine_pvalues_discrete` implements this approach in a fast, thorough, and tested manner...", I can confirm "fast" and "tested", but what is meant by "thorough" in this case?
- [x] 3. Similarly, can you explain "taking care of pitfalls such as correctly handling complements, sidedness, and empirical p-values"? By "empirical p-values", you mean the discrete distribution of p-values instead of assuming the continuous uniform distribution (the core idea of the package)?
- [x] 4. Toward the end of the paper, you wrote "This approach is considerably faster than a permutation test starting at the level of individual datasets". I can imagine things that this *could* mean, but I'm not certain of what *you* meant. Can you expand on the permutation test idea?
- [x] 5. I think I understood most of the last two sentences of the last paragraph, but I don't understand how they related to the first sentence. "Note that this module de-emphasises some typical applications of combining 𝑝 values as it can be rarely applied to them." Can you explain?
- [x] 6. One part of those sentences I didn't understand was "the 𝑝 values to be combined often originate from continuous tests or tests with a discrete but rather dense supports, and one rarely has the luxury to know the null distribution of 𝑝 values". I assume that when you write the _test_ is "continuous" or "discrete", you are ultimately referring to whether the _distribution of p-values under the null hypothesis_ is continuous or discrete. If the p-values originated from "continuous tests", and the tests are valid (e.g. "size" of the test is nominal), isn't the distribution of p-values under the null hypothesis the continuous uniform distribution (by definition)? If the tests are discrete but have "rather dense supports", arent the distributions well-approximated by the continuous uniform distribution? So these are the cases when we *do* have the luxury of knowing the null distribution of p-values, no? Your package *can* be applied to these cases, but it is more general because it *also* handles the case of when the distribution of p-values under the null hypothesis is (very) discrete.
- [x] 7. For the "community guidelines" criterion: I see a link to the issue tracker from the main repo page, but I don't see guidelines for contributing or seeking support. I'd suggest making these more prominent so I can check that box.
- [x] 8. For addressing the "state of the field" criterion, it would be ideal (although not required) if you could establish that no popular packages do what this package does. If that's true, I think the best way to demonstrate it would be to name several other (beyond the two already cited) closely related packages or routines and mention that they are all for continuous tests. If it's not true, just explain why this package is also needed (e.g. there is not one in Python, or there is not one that is permissively-licensed, or the existing ones are poorly documented).
