### 2023-11-30 7341af9
Clarifying some passages, addressing #1.
