+++ paper.bib
@@ -47,3 +47,15 @@
 	url = {https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.combine_pvalues.html},
 }
 
+@article{Phipson2010,
+	url = {https://doi.org/10.2202/1544-6115.1585},
+	title = {Permutation $P$-values Should Never Be Zero: Calculating Exact $P$-values When Permutations Are Randomly Drawn},
+	author = {Belinda Phipson and Gordon K Smyth},
+	volume = {9},
+	number = {1},
+	pages = {39},
+	journal = {Statistical Applications in Genetics and Molecular Biology},
+	doi = {10.2202/1544-6115.1585},
+	year = {2010},
+}
+
