+++ paper.md
@@ -41,11 +41,13 @@ Discrete tests particularly include all rank tests, which are an appropriate cho
 Thanks to modern computing, we can solve this problem using simple Monte Carlo simulations:
 For each test, we sample one $p$ value from each of the respective discrete null distributions, apply the combining statistics to these, and repeat this until we obtain a good estimate of the null distribution of the combining statistics.
 Finally, we compare the combining statistics of the actual data to estimate the combined $p$ value.
-The presented module `combine_pvalues_discrete` implements this approach in a fast, thorough, and tested manner, taking care of pitfalls such as correctly handling complements, sidedness, and empirical $p$ values as well as handling tedious and error-prone tasks such as determining the null distribution $p$ values for a given test and sample size.
-This approach is considerably faster than a permutation test starting at the level of individual datasets.
+The presented module `combine_pvalues_discrete` implements this approach in a fast, thorough, and tested manner:
+It takes care of pitfalls such as correctly handling complements, sidedness, and $p$ values from a Monte Carlo estimate of the null distribution [@Phipson2010].
+Also, it handles tedious and error-prone tasks such as determining the null distribution $p$ values for a given test and sample size.
+This approach is considerably faster than a permutation test starting at the level of sub-datasets, i.e., a Monte Carlo null model of the entire dataset obtained by combining samples from Monte Carlo null models of the sub-datasets.
 
-Note that this module de-emphasises some typical applications of combining $p$ values as it can be rarely applied to them.
-For example, when performing meta analyses, the $p$ values to be combined often originate from continuous tests or tests with a discrete but rather dense supports, and one rarely has the luxury to know the null distribution of $p$ values.
+Note that this module de-emphasises some typical applications of combining $p$ values in which we expect known discrete null distributions of $p$ values to rarely occur.
+For example, when performing meta analyses, the null distributions of $p$ values are either continuous, close to it, or not known.
 However, as a side product, our module contains weighted versions of popular combining methods that may be of interest to researchers combining continuous tests.
 
 # Acknowledgements
