+++ README.rst
@@ -7,5 +7,6 @@ A Python module for combining *p* values of discrete tests.
   * or similar.
   
 * `Documentation <https://combine-p-values-discrete.rtfd.io>`_
-* `Issue Tracker <https://github.com/BPSB/combine-p-values-discrete/issues>`_ – Please report any bugs here and feel free to request features (in particular support for tests).
+* `Issue Tracker <https://github.com/BPSB/combine-p-values-discrete/issues>`_ – Please use this if you need any help with this module, in particular if you are facing any bugs or require features (in particular support for tests).
+* Contributions are welcome. Feel free to discuss them on the issue tracker or just file a pull request.
 * DOI for citing: `10.5281/zenodo.8338798 <https://dx.doi.org/10.5281/zenodo.8338798>`_
