+++ paper.bib
@@ -59,3 +59,17 @@
 	year = {2010},
 }
 
+@misc{PyMare,
+	author = {Yarkoni T, Salo T, Nichols T, Peraza J.},
+	title = {PyMARE: Python Meta-Analysis \& Regression Engine},
+	doi = {https://zenodo.org/doi/10.5281/zenodo.4266738},
+	year = {2022},
+	url = {https://pymare.readthedocs.io}
+}
+
+@misc{MetaP,
+	title = {metap: Meta-Analysis of Significance Values},
+	author = {Michael Dewey},
+	url = {https://cran.r-project.org/package=metap},
+}
+
