+++ paper.bib
@@ -49,7 +49,7 @@
 
 @article{Phipson2010,
 	url = {https://doi.org/10.2202/1544-6115.1585},
-	title = {Permutation $P$-values Should Never Be Zero: Calculating Exact $P$-values When Permutations Are Randomly Drawn},
+	title = {Permutation $p$-values Should Never Be Zero: Calculating Exact $p$-values When Permutations Are Randomly Drawn},
 	author = {Belinda Phipson and Gordon K Smyth},
 	volume = {9},
 	number = {1},
@@ -60,7 +60,7 @@
 }
 
 @misc{PyMare,
-	author = {Yarkoni T, Salo T, Nichols T, Peraza J.},
+	author = {Yarkoni T and Salo T and Nichols T and Peraza J},
 	title = {PyMARE: Python Meta-Analysis \& Regression Engine},
 	doi = {https://zenodo.org/doi/10.5281/zenodo.4266738},
 	year = {2022},
