+++ paper.bib
@@ -60,9 +60,9 @@
 }
 
 @misc{PyMare,
-	author = {Yarkoni T and Salo T and Nichols T and Peraza J},
+	author = {Tal Yarkoni and Taylor Salo and Julio A. Peraza and Thomas E. Nichols},
 	title = {PyMARE: Python Meta-Analysis \& Regression Engine},
-	doi = {https://zenodo.org/doi/10.5281/zenodo.4266738},
+	doi = {10.5281/zenodo.4266738},
 	year = {2022},
 	url = {https://pymare.readthedocs.io}
 }
