### 2023-12-01 BPSB/combine-p-values-discrete#1 @Wrzlprmft
> Can you respond to JOSS's criteria for "substantial scholarly effort", specifically "As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual."? It's unclear from the commit history whether that's satisfied, but commit histories don't tell the full story. (I know because I submitted a manuscript to JOSS for which this was questioned. Of course, I wouldn't have submitted if I didn't believe the criterion had been satisfied. But as a reviewer, I need to ask!)

[Your guesses](https://github.com/openjournals/joss-reviews/issues/6096#issuecomment-1830855128) were mostly accurate. In early 2022, I stumbled upon this problem in my own research (which is unfortunately still unpublished and not accessible enough for an application example) and had the luxury of investing roughly a month into this almost full time. This includes explorations of the problem and possible solutions, which are not reflected in the commit history. I then revisited the project twice (after a month and two months ago) for roughly a week to check, revise, and extend it with a fresh eye.

Without tallying my work time on this, there is no denying that I do not comply with the cited rule of thumb. I submitted despite this in confidence that my work complies with the spirit of the rule for the following reason: The reason this module required so little work from me was that the required background knowledge and skills overlapped quite well with my experience, both on statistics and software development. Had I engaged in the same project about seven years ago, I would have easily complied with the rule of thumb by spending considerably more time on literature research, understanding problems, making things efficient, devising tests, fighting bugs, etc. I would expect a similar outcome if I had assigned this to a student as a thesis project or similar. Sampling other works published in JOSS, I considered this module to be en par with many of them.

Now, I cannot expect you to trust my word on this, but I can also hardly substantiate this by means other than the material you are already reviewing. I can just hope that you agree with my self-assessment.

> In the claim "The presented `module combine_pvalues_discrete` implements this approach in a fast, thorough, and tested manner...", I can confirm "fast" and "tested", but what is meant by "thorough" in this case?

The intended meaning is more or less the opposite of what JOSS calls a “half-baked solution”: Edge cases are covered; relevant options and features are available; the things mentioned in the second half of the sentence. I concede that this may be somewhat vague, boastful, and possibly redundant. On the other hand, this summarises aspects that make up a considerable part of work and why users should employ this module instead of a home-made solution. I exemplify this point to some extent in the following, but I feel that a summary may be relevant to readers who are less familiar with the problem at hand.

> Similarly, can you explain "taking care of pitfalls such as correctly handling complements, sidedness, and empirical p-values"? By "empirical p-values", you mean the discrete distribution of p-values instead of assuming the continuous uniform distribution (the core idea of the package)?

I am referring to _p_ values obtained by counting, i.e., using a Monte Carlo estimate of the null distribution.

I now write “[handling] $p$ values from a Monte Carlo estimate of the null distribution” and split up the respective sentence, which was very long anyway. I also cite the paper by Phipson and Smyth you referred to on this.

> Toward the end of the paper, you wrote "This approach is considerably faster than a permutation test starting at the level of individual datasets". I can imagine things that this _could_ mean, but I'm not certain of what _you_ meant. Can you expand on the permutation test idea?

This is about simulating a null model for each sub-dataset (usually a permutation test) and then combining the results. This is what is done in [Checking the result](https://combine-p-values-discrete.readthedocs.io/en/latest/#checking-the-result) in the documentation and also what I assume you do when you check that “when the null hypothesis of each individual test is true, the distribution of combined p-values […] appears uniform”.

Trying to clarify this, I now write: “This approach is considerably faster than a permutation test starting at the level of sub-datasets, i.e., a Monte Carlo null model of the entire dataset obtained by combining samples from Monte Carlo null models of the sub-datasets.”

> I think I understood most of the last two sentences of the last paragraph, but I don't understand how they related to the first sentence. "Note that this module de-emphasises some typical applications of combining 𝑝 values as it can be rarely applied to them." Can you explain?
>
> […]
>
>  One part of those sentences I didn't understand was "the 𝑝 values to be combined often originate from continuous tests or tests with a discrete but rather dense supports, and one rarely has the luxury to know the null distribution of 𝑝 values". I assume that when you write the _test_ is "continuous" or "discrete", you are ultimately referring to whether the _distribution of p-values under the null hypothesis_ is continuous or discrete. If the p-values originated from "continuous tests", and the tests are valid (e.g. "size" of the test is nominal), isn't the distribution of p-values under the null hypothesis the continuous uniform distribution (by definition)? If the tests are discrete but have "rather dense supports", arent the distributions well-approximated by the continuous uniform distribution? So these are the cases when we _do_ have the luxury of knowing the null distribution of p-values, no? Your package _can_ be applied to these cases, but it is more general because it _also_ handles the case of when the distribution of p-values under the null hypothesis is (very) discrete.

I now see how the logic of those sentences was broken or not clear.

The point of the sentences in question is this:
A considerable amount of literature on combining _p_ values focuses on meta analyses and specific methods for those. However, I admit that my module is limited in this respect, simply because I expect that many meta analyses either lack basic information required for benefitting from this module (the null distribution of _p_ values) or do not suffer from the problem solved by this module (discrete null distributions). The sentences in question delineate the module’s scope in this direction and similar ones.

While this module can indeed be applied to continuous tests, I expect that most users performing a meta analysis on such tests will find other software cater better to their needs, either because it has features specific to meta analyses or does not add unneeded complexity.

I now write: “Note that this module de-emphasises some typical applications of combining $p$ values in which we expect known discrete null distributions of $p$ values to rarely occur. For example, when performing meta analyses, the null distributions of $p$ values are either continuous, close to it, or not known. However, as a side product, our module contains weighted versions of popular combining methods that may be of interest to researchers combining continuous tests.”

> For the "community guidelines" criterion: I see a link to the issue tracker from the main repo page, but I don't see guidelines for contributing or seeking support. I'd suggest making these more prominent so I can check that box.

I made clear that the issue tracker may also be used for support and offering contributions (as well as straight pull requests).

> For addressing the "state of the field" criterion, it would be ideal (although not required) if you could establish that no popular packages do what this package does. If that's true, I think the best way to demonstrate it would be to name several other (beyond the two already cited) closely related packages or routines and mention that they are all for continuous tests. If it's not true, just explain why this package is also needed (e.g. there is not one in Python, or there is not one that is permissively-licensed, or the existing ones are poorly documented).

I have looked into the modules named by you and some others:

* PythonMeta apparently doesn’t combine _p_ values at all. However, its documentation is not very expansive.

* PyMARE admits itself that it is mostly a service package for NiMARE. It supports Fisher’s and Stouffer’s method. In my understanding it does combine _p_ values by some permutation test, but I see no way to specify the null distribution of _p_ values of a dataset, so I assume these aren’t accounted for.

* Statsmodels does not feature methods to combine _p_ values at all. At least, none of the names of popular methods appear in its source (except _Fisher_ of course, but only in different contexts).

* The R packages `poolr` and `metap` feature methods for combining _p_ values, but their documentation does not specify a way to specify the null distribution of _p_values.

* The R packages `metafor`, `meta`, `rmeta` do not feature methods to combine _p_ values. At least I cannot find any promiment method names in the documentation.

* There appears to be a basic Matlab implementation of Fisher’s method, which takes an array of  _p_ values as the only argument.

I now cite all Python and R packages that at least feature some method for combining _p_ values. I refrained from citing the others, as they are not even competing in the same arena. I also explicitly mention that I am aware of no software solution to the problem at hand.
