### 2023-12-02 BPSB/combine-p-values-discrete#1 @mdhaber
1. I agree that is is substantial and scholarly.
2. I see what "thorough" means now. I don't think it's boastful because I can easily imagine a half-baked solution, and this is not that. I don't mind leaving it vague, since I can't find one word that would do a better job of summing up the intent. If it is important enough, you could elaborate, but I am not pushing for it.
3. OK, I think that is an improvement. Computing p-values from Monte Carlo simulation is definitely error-prone, as the paper points out. IIUC, "[Complements"](https://combine-p-values-discrete.readthedocs.io/en/latest/#complements) explains the potential "complements" pitfall, and the `alternative` parameter documentation of `[combine`](https://combine-p-values-discrete.readthedocs.io/en/latest/#combine_pvalues_discrete.combine) explains the "sidedness" pitfall?
4. "This is about simulating a null model for each sub-dataset (usually a permutation test)"
I think I see the misunderstanding.

IIUC your approach starts with the (discrete) null distributions of p-values of each test.

This is much faster than the simplest approach - (effectively) drawing from the distribution of p-values of each test by drawing random data that is consistent with the null hypothesis of each test, and explicitly performing each test on the random data. _This_ is what you do in ["Checking the results"](https://combine-p-values-discrete.readthedocs.io/en/latest/#checking-the-result) and what I did using `scipy.stats.monte_carlo_test` below:

```python3
from functools import reduce
import numpy as np
from scipy import stats

rng = np.random.default_rng(38345834152534)

data = [# control group  , treatment group
        ( [20,44,14,68]  , [73,22,80]    ), # Breed 1
        ( [60,68,46,45]  , [30]          ), # Breed 2
        ( [92,97,98]     , [84,89]       ), # …
        ( [14]           , [21,45,31,23] ),
        ( [24,58,0,24,33], [65,51,61]    ),
        ( [93,76,70,83]  , [84]          ),
        ( [10,2]         , [28,36,11]    ),
        ( [27]           , [38,58]       ),
        ( [8,13,37]      , [43,51]       ),
        ( [18]           , [12]          ),
]
data = reduce(lambda x, y: x + y, data)  # reshape data

def fisher_statistic(*data, axis=-1):
  pvalues = [stats.mannwhitneyu(data[i], data[i+1], axis=axis, alternative="less").pvalue
             for i in range(0, len(data), 2)]
  return -2*np.sum(np.log(pvalues), axis=0)

res = stats.monte_carlo_test(data, [rng.random]*len(data), fisher_statistic, alternative='greater')
res.pvalue
```
A modification of this second approach would be to (effectively) draw from the distribution of p-values of each test by randomly "permuting" the test data in the manner consistent with the null hypothesis of each test, and explicitly performing each test on its permuted data.

  Maybe it would be more clear to me if you were to leave out the bit about a "permuation test" and emphasize that your Monte Carlo test draws directly from the null distributions of p-values of each test rather than indirectly, by performing the test on data simulated under the test's null hypothesis.

5. (and 6.) Here's what I'm taking from this: This package does not claim to be the best choice for *all* meta-analyses. For example, in some meta-analyses, the complete null distributions cannot be generated efficiently, so the approach taken by this package doesn't apply. (Instead, one would need to sample "indirectly" from the null distribution of p-values as described above, by performing the test on data simulated under the test's null hypothesis.) In other meta-analyses, the null distributions of the individual tests are either continuous or nearly so; in such cases, `combine_pvalues` or similar may be sufficient. Even in those cases, however, this package could be beneficial if the user wanted to add weights.
Is this similar to what you wanted me to get from the paragraph?

7. Great. I've checked the box.
8. Nice, I think that addresses this point.
