### 2023-12-02 BPSB/combine-p-values-discrete#1 @Wrzlprmft
> 2. I see what "thorough" means now. I don't think it's boastful because I can easily imagine a half-baked solution, and this is not that. I don't mind leaving it vague, since I can't find one word that would do a better job of summing up the intent. If it is important enough, you could elaborate, but I am not pushing for it.

I was also browsing a thesaurus for a better word. I would consider the following sentences the elaboration, as indicated by the colon following “fast, thorough, and tested manner”. I don’t see another way to elaborate this without being overly wordy or clunky.

> 3. OK, I think that is an improvement. Computing p-values from Monte Carlo simulation is definitely error-prone, as the paper points out. IIUC, "[Complements"](https://combine-p-values-discrete.readthedocs.io/en/latest/#complements) explains the potential "complements" pitfall, and the `alternative` parameter documentation of `[combine`](https://combine-p-values-discrete.readthedocs.io/en/latest/#combine_pvalues_discrete.combine) explains the "sidedness" pitfall?

All correctly identified.

> 4. "This is about simulating a null model for each sub-dataset (usually a permutation test)"
>        I think I see the misunderstanding.
>
> IIUC your approach starts with the (discrete) null distributions of p-values of each test.

Correct.

> This is much faster than the simplest approach - (effectively) drawing from the distribution of p-values of each test by drawing random data that is consistent with the null hypothesis of each test, and explicitly performing each test on the random data. […] A modification of this second approach would be to (effectively) draw from the distribution of p-values of each test by randomly "permuting" the test data in the manner consistent with the null hypothesis of each test, and explicitly performing each test on its permuted data.

I see your point. As the two approaches are equivalent (for ranked tests without ties), I had them both filed as permutation test. I started with the second approach and then immediately noticed that permuting the actual data just complicates things.

> Maybe it would be more clear to me if you were to leave out the bit about a "permuation test" and emphasize that your Monte Carlo test draws directly from the null distributions of p-values of each test rather than indirectly, by performing the test on data simulated under the test's null hypothesis.

I see. I now write: “By storing and sampling the null distribution of $p$ values for each sub-dataset, our approach is considerably faster than using a Monte Carlo null model for the entire dataset by combining samples from Monte Carlo null models of the sub-datasets.”

> 6. Here's what I'm taking from this: This package does not claim to be the best choice for _all_ meta-analyses. For example, in some meta-analyses, the complete null distributions cannot be generated efficiently, so the approach taken by this package doesn't apply. (Instead, one would need to sample "indirectly" from the null distribution of p-values as described above, by performing the test on data simulated under the test's null hypothesis.) In other meta-analyses, the null distributions of the individual tests are either continuous or nearly so; in such cases, `combine_pvalues` or similar may be sufficient. Even in those cases, however, this package could be beneficial if the user wanted to add weights.
>        Is this similar to what you wanted me to get from the paragraph?

Mostly (if I understand correctly).

A potential source of misunderstanding may be what counts as a _meta analysis_ and whether there are other applications of combining _p_ values. My understanding of the term is that if I perform a meta analysis, I combine the results of existing (usually published) studies, and thus I have limited access to the data, the studies usually already have a large sample size (and thus the null distribution of _p_ values is close to continuous, even if the test is discrete), etc. The dog example would not be a meta analysis in this sense. Since there may be other understandings of the term, I now try to disambiguate by writing “a meta analysis of existing studies” instead of just “meta analyses”.

Moreover, meta analysis goes much further than just combining _p_ values, considering effect sizes, dependencies, publication bias, etc. For instance, in the dog example, one would report the median fold increase to quantify the effect of the drug, and the only question (addressed by this module) would be whether this is a product of chance. In many meta studies, even providing such a summarising quantity is challenging. As mentioned above, there are huge modules dedicated to meta analysis that do not even contain methods for combining _p_ values.

Most of the existing modules for combining _p_ values are aimed at meta analyses (in the sense of the term I use above) and have specific features dedicated to such applications, which in turn are missing from this module. The paragraph also serves to explain this and throttle respective expectations.
