### 2023-12-02 e412460
Further clarifications, addressing Issue #1.
