+++ paper.md
@@ -45,10 +45,10 @@ Finally, we compare the combining statistics of the actual data to estimate the 
 The presented module `combine_pvalues_discrete` implements this approach in a fast, thorough, and tested manner:
 It takes care of pitfalls such as correctly handling complements, sidedness, and $p$ values from a Monte Carlo estimate of the null distribution [@Phipson2010].
 Also, it handles tedious and error-prone tasks such as determining the null distribution $p$ values for a given test and sample size.
-This approach is considerably faster than a permutation test starting at the level of sub-datasets, i.e., a Monte Carlo null model of the entire dataset obtained by combining samples from Monte Carlo null models of the sub-datasets.
+By storing and sampling the null distribution of $p$ values for each sub-dataset, our approach is considerably faster than using a Monte Carlo null model for the entire dataset by combining samples from Monte Carlo null models of the sub-datasets.
 
 Note that this module de-emphasises some typical applications of combining $p$ values in which we expect known discrete null distributions of $p$ values to rarely occur.
-For example, when performing meta analyses, the null distributions of $p$ values are either continuous, close to it, or not known.
+For example, when performing a meta analysis of existing studies, the null distributions of $p$ values are either continuous, close to it, or not known.
 However, as a side product, our module contains weighted versions of popular combining methods that may be of interest to researchers combining continuous tests.
 
 # Acknowledgements
