+++ docs/index.rst
@@ -99,7 +99,7 @@ These objects can then be combined using the `combine` function.
 
 The difficulty for determining the combined *p* value is convolving the respective null distributions.
 While this is analytically possible for continuous tests or a small number of discrete tests, it requires numerical approximations otherwise due to a combinatorial explosion (e.g., even for the small dataset in `comparison`, we we would have to handle 53508000 combinations).
-To perform these approximations, we use a Monte Carlo sampling of combinations of individual *p* values.
+To perform these approximations, we use a Monte Carlo sampling of combinations of individual *p* values (from the respective null distributions).
 Thanks to modern computing and NumPy, it is easy to make the number of samples very high and the result very accurate.
 
 .. _complements:
