### 2024-01-26 #6096 @Wrzlprmft
> Hi, @Wrzlprmft could you please update us on your status in addressing issues raised by @mdhaber?

In my understanding, there is nothing for me to address at the moment. I addressed @mdhaber’s last comments [here](https://github.com/BPSB/combine-p-values-discrete/issues/1) (and the commits referenced therein). Since my last reply was mostly asserting his understandings, I didn’t expect a response and assumed no further clarification was needed.
