+++ paper.bib
@@ -68,7 +68,7 @@
 }
 
 @misc{MetaP,
-	title = {metap: Meta-Analysis of Significance Values},
+	title = {{metap}: Meta-Analysis of Significance Values},
 	author = {Michael Dewey},
 	url = {https://cran.r-project.org/package=metap},
 }
