### 2024-02-03 #6096 @Wrzlprmft
> References look good except for "metap" probably isn't intended to be capitalized in the rendered document.

Fixed. While I disagree with lowercasing proper names, this is reproducing the lack of capitalisation of the cited source.
