### 2024-02-18 bf24c97
Adding examples to the public interface, addressing Issue #3.
