### 2024-02-18 da9a519
Suggestions from Issue #2 that might improve CI and user installation.
