+++ requirements.txt
@@ -1,3 +1,2 @@
 numpy
 scipy>=1.8.1
-statsmodels
