+++ .github/workflows/test.yml
@@ -14,7 +14,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pytest pytest-rng
+          pip install .[test]
           pip install .
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with pytest
