### 2024-02-18 1ee423b
Documenting how to run tests, addressing Issue #3
