+++ README.rst
@@ -5,7 +5,8 @@ A Python module for combining *p* values of discrete tests.
   * ``pip install combine-pvalues-discrete``
   * ``pip3 install git+git://github.com/BPSB/combine-p-values-discrete``
   * or similar.
-  
+
+* Install requirements for testing with `pip install ".[test]"` (or similar). Run the test with `pytest` (or similar) in the source directory. Run `pytest -m "not slow"` to exclude some particularly time-consuming tests.
 * `Documentation <https://combine-p-values-discrete.rtfd.io>`_
 * `Issue Tracker <https://github.com/BPSB/combine-p-values-discrete/issues>`_ – Please use this if you need any help with this module, in particular if you are facing any bugs or require features (in particular support for tests).
 * Contributions are welcome. Feel free to discuss them on the issue tracker or just file a pull request.
