### 2024-02-20 #6096 @Wrzlprmft
> @Wrzlprmft it seems that both reviewers recommend acceptance!

So it does.

Thanks to the reviewers (@mdhaber and @steppi) for their diligent criticism and constructive comments.

> When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (see this guide). Please do this and post the version number and archive DOI here.

* Version: 1.2.2
* DOI: 10.5281/zenodo.10684537
