### 2024-02-21 #6096 @Wrzlprmft
> @Wrzlprmft please edit the zenodo record so that the title, author names and affiliation are the same as in paper.

Hopefully done. I added the paper title as an alternative title. Please tell me if it should be the main title.
