### 2024-02-22 #6096 @Wrzlprmft
> @Wrzlprmft please change the main title of zenodo to "A Python module to combine 𝑝 values arising from1 discrete tests." i.e. as in JOSS paper so that the DOI looks correct and has the same title as in your paper.
>
> https://zenodo.org/records/10684537

Done (except for the stray “1” which I presume to be an error).
